requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ChatterBot==1.0.2
chatterbot-corpus==1.2.0
click==8.1.7
codegen==1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
DateTime==5.5
debugpy==1.8.6
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
Django==5.0.4
django-dotenv==1.4.2
dnspython==2.7.0
executing==2.1.0
favicon==0.7.0
filelock==3.13.4
Flask==3.0.2
Flask-Caching==2.3.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.46.0
frozenlist==1.4.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
googlesearch-python==1.2.5
greenlet==3.1.1
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imbalanced-learn==0.12.3
imblearn==0.0
importlib-metadata==7.0.1
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.2.1
kiwisolver==1.4.5
libclang==18.1.1
lightgbm==4.5.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mathparse==0.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
multidict==6.0.5
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysqlclient==2.2.4
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.2
openai==0.28.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parso==0.8.4
patterns==0.3
Pillow==10.1.0
Pint==0.23
platformdirs==4.2.0
prompt_toolkit==3.0.48
proto-plus==1.23.0
protobuf==4.21.12
psutil==6.0.0
pure_eval==0.2.3
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.17.2
pymongo==3.13.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.4
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
pywin32==306
PyYAML==3.13
pyzmq==26.2.0
random2==1.0.1
referencing==0.33.0
regex==2024.4.28
requests==2.31.0
requests-file==2.1.0
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
selenium==4.17.2
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.5.0
stack-data==0.6.3
streamlit==1.30.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.2.0
tldextract==5.1.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.0
utils==1.0.2
validators==0.22.0
virtualenv==20.25.3
virtualenvwrapper-win==1.2.7
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.43.0
whois==1.20240129.2
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.17.0
zope.interface==7.0.3
